Move all traits to one module
Easier to know what traits are needed if they are all together.
Also makes management of the Result/Error types easier.
